engine: extract dispatch::elimination (#159)
